requirements.txt
django
djangorestframework
djangorestframework-simplejwt
django-cors-headers
phonenumbers
django-phonenumber-field
python-dotenv
pillow
drf-spectacular
pandas
openpyxl
